requirements.txt
django-cors-headers>=3.7
factory_boy==2.12.0
gunicorn[gevent]==20.0.4
healpy>=1.14
psycopg2===2.8.4
tom-alerts-dash[scimma]>=0.1.2
tom-superevents>=0.1
django-webpack-loader==0.6.0
tomtoolkit>=2.6.5
whitenoise~=5.2
tom-antares
tom-fink
antares-client==1.1
fink-client~=2.0
tom-pittgoogle==0.1.11
hop-client==0.5.0
avro==1.10.2
